[wip] wcxb: Add int_mode=3 which allows a high res timer to drive interrupt processing.
This patch is not ready for primetime since it only compiles against CentOS 5.
It will need some more changes to make it work against newer kernels.

For systems that do not route interrupts properly, this will allow indicating
that we want the interrupt handling routines to be checked from the context of a
high resolution timer.

Internal-Issue-ID: DAHDI-1033
Signed-off-by: Shaun Ruffell <sruffell@digium.com>